fix: knowledge base settings (LLM-001)
