+++ packages/base-types/src/models/project/index.ts
@@ -1,11 +1,13 @@
 import { AnyRecord } from '@voiceflow/common';
 
 import { LinkType } from '../base/link';
+import { KnowledgeBase } from './knowledgeBase';
 import { Member } from './member';
 import { Prototype } from './prototype';
 import { Themes } from './theme';
 
 export { LinkType } from '../base/link';
+export * from './knowledgeBase';
 export * from './member';
 export * from './prototype';
 export * from './theme';
@@ -30,52 +32,6 @@ export interface AIAssistSettings {
   generateStep?: boolean;
   generateNoMatch?: boolean;
   generativeTasks?: boolean;
-}
-
-export enum KnowledgeBaseDocumentType {
-  PDF = 'pdf',
-  TEXT = 'text',
-  URL = 'url',
-}
-
-export interface KnowledgeBaseData {
-  type: KnowledgeBaseDocumentType;
-  name: string;
-}
-
-export interface KnowledgeBasePDF extends KnowledgeBaseData {
-  type: KnowledgeBaseDocumentType.PDF;
-  name: string;
-}
-
-export interface KnowledgeBaseText extends KnowledgeBaseData {
-  type: KnowledgeBaseDocumentType.TEXT;
-  name: string;
-}
-
-export interface KnowledgeBaseURL extends KnowledgeBaseData {
-  type: KnowledgeBaseDocumentType.URL;
-  name: string;
-  url: string;
-}
-
-export enum KnowledgeBaseDocumentStatus {
-  ERROR = 'ERROR',
-  SUCCESS = 'SUCCESS',
-  PENDING = 'PENDING',
-  INITIALIZED = 'INITIALIZED',
-}
-
-export interface KnowledgeBaseDocument {
-  data: KnowledgeBasePDF | KnowledgeBaseText | KnowledgeBaseURL;
-  status: {
-    type: KnowledgeBaseDocumentStatus;
-    data?: unknown;
-  };
-  creatorID: number;
-  updatedAt: Date;
-  documentID: string;
-  s3ObjectRef: string;
 }
 
 export interface Model<PlatformData extends AnyRecord, MemberPlatformData extends AnyRecord> {
@@ -104,5 +60,5 @@ export interface Model<PlatformData extends AnyRecord, MemberPlatformData extend
   members: Member<MemberPlatformData>[];
   platformData: PlatformData;
 
-  knowledgeBase?: Record<string, KnowledgeBaseDocument>;
+  knowledgeBase?: KnowledgeBase;
 }
